+++ src/routes.tsx
@@ -4,8 +4,7 @@ import PrivateLayout from './components/common/layouts/PrivateLayout';
 import routes from './constants/routes';
 import ChallengDetailPage from './pages/ChallengDetailPage';
 import ChallengeListPage from './pages/ChallengeListPage';
-import FamilyCreatePage from './pages/FamilyCreatePage/FamilyCreatePage';
-import FamilyEntryPage from './pages/FamilyEntryPage';
+import FamilyCreatePage from './pages/FamilyCreatePage';
 import LoginPage from './pages/LoginPage';
 import MainPage from './pages/MainPage/MainPage';
 import MyPage from './pages/MyPage';
